requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
ansi2html=1.9.1=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
brotli-python=1.0.9=py310hd8f1fbe_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.11.17=hbcca054_0
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
comm=0.2.1=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.0=pypi_0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.0=0
cuda-compiler=11.6.0=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.55=h9dd2d0c_0
cuda-cupti=11.6.55=h43c5c43_0
cuda-cuxxfilt=11.6.55=h69de05d_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.6.55=hff0b7d4_0
cuda-libraries=11.6.0=0
cuda-libraries-dev=11.6.0=0
cuda-memcheck=11.6.55=h0288dce_0
cuda-nsight=11.6.55=0
cuda-nsight-compute=11.6.0=0
cuda-nvcc=11.6.55=h5758ece_0
cuda-nvdisasm=11.6.55=h5556c0d_0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.6.55=h30b2dac_0
cuda-nvprune=11.6.55=h3791f62_0
cuda-nvrtc=11.6.55=hc54fff9_0
cuda-nvrtc-dev=11.6.55=h42ad0f4_0
cuda-nvtx=11.6.55=h99d0529_0
cuda-nvvp=11.6.58=h67ee751_0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.6.55=h4716e2e_0
cuda-toolkit=11.6.0=0
cuda-tools=11.6.0=0
cuda-visual-tools=11.6.0=0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.12.1=pypi_0
cyclonedds=0.10.2=pypi_0
dash=2.14.2=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
decorator=5.1.1=pypi_0
diff-gaussian-rasterization=0.0.0=pypi_0
docker-pycreds=0.4.0=pypi_0
evaluate-3d-reconstruction=0.1.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
flask=3.0.0=pypi_0
fonttools=4.47.0=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2023.12.2=pypi_0
gds-tools=1.2.0.100=0
giflib=5.2.1=h36c2ea0_2
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
huggingface-hub=0.20.2=pypi_0
idna=3.6=pyhd8ed1ab_0
imageio=2.33.1=pypi_0
importlib-metadata=7.0.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.19.0=pypi_0
ipywidgets=8.1.1=pypi_0
itsdangerous=2.1.2=pypi_0
jbig=2.1=h7f98852_2003
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h166bdaf_1
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.6.1=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.1=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
lerc=2.2.1=h9c3ff4c_0
libcublas=11.8.1.74=h1e58c10_0
libcublas-dev=11.8.1.74=h7a51e1f_0
libcufft=10.7.0.55=h563f203_0
libcufft-dev=10.7.0.55=h05eb8d0_0
libcufile=1.2.0.100=0
libcufile-dev=1.2.0.100=0
libcurand=10.2.9.55=h7c349da_0
libcurand-dev=10.2.9.55=hd2e71f0_0
libcusolver=11.3.2.55=hebb49eb_0
libcusparse=11.7.1.55=h9a152cf_0
libdeflate=1.7=h7f98852_5
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libnpp=11.6.0.55=hdb0c674_0
libnpp-dev=11.6.0.55=h0163868_0
libnvjpeg=11.6.0.55=h6f17e28_0
libnvjpeg-dev=11.6.0.55=h0163868_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_1
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
lightning-utilities=0.10.0=pypi_0
lpips=0.1.4=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310ha2c4b55_0
mkl_fft=1.3.1=py310h2b4bcf5_1
mkl_random=1.2.2=py310h00e6091_0
natsort=8.4.0=pypi_0
nbformat=5.5.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
nettle=3.6=he412f7d_0
nsight-compute=2022.1.0.12=0
numpy=1.24.3=py310hd5efca6_0
numpy-base=1.24.3=py310h8e6c178_0
open3d=0.16.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h780b84a_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pillow=9.4.0=py310h6a678d5_0
pip=23.3.1=py310h06a4308_0
platformdirs=4.1.0=pypi_0
plotly=5.18.0=pypi_0
plyfile=1.0.3=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.7=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda11.6_cudnn8.3.2_0
pytorch-msssim=1.0.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.32.0=pypi_0
requests=2.31.0=pyhd8ed1ab_0
retrying=1.3.4=pypi_0
rich=13.7.0=pypi_0
rich-click=1.7.2=pypi_0
rpds-py=0.16.2=pypi_0
safetensors=0.4.1=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
sentry-sdk=1.39.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py310_cu116
torchmetrics=1.2.1=pypi_0
torchvision=0.13.1=py310_cu116
tqdm=4.65.0=pypi_0
traitlets=5.14.1=pypi_0
trimesh=4.4.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
urllib3=2.1.0=pyhd8ed1ab_0
wandb=0.16.1=pypi_0
wcwidth=0.2.12=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.0=ha95c52a_0
